corectie minora pentru Test_Vector
(cherry picked from commit 77052e6d8b9859a91cedb5b8934cfa5a9fd8a637)

+++ PrezicerePretActiuni/Test_Vector.cpp
@@ -1,11 +1,11 @@
 #include "Test_Vector.h"
 
-int Test_Vector::get_Dim_Test()
+size_t Test_Vector::get_Dim_Test()
 {
 	return this->test.size();
 }
 
-int Test_Vector::get_Dim_Rezultat()
+size_t Test_Vector::get_Dim_Rezultat()
 {
 	return this->rezultat.size();
 }
